Put popup into own component. The large screen layout now uses a portal based popup in an overlay over the sidebar. The small screen layout still uses the autocomplete within the fullsize address input
