+++ src/App.module.css
@@ -27,7 +27,7 @@
 
 .popupContainer {
     grid-column: 1 / span 1;
-    grid-row: 1 /span 1;
+    grid-row: 1 / span 1;
     background-color: transparent;
     pointer-events: none;
     z-index: 2;
