+++ src/sidebar/search/PopUp.tsx
@@ -1,0 +1,54 @@
+import { ReactNode, useEffect, useRef } from 'react'
+import { createPortal } from 'react-dom'
+import styles from './PopUp.module.css'
+
+export interface PopUpProps {
+    children: ReactNode
+    rect: DOMRect
+    keepClearAtBottom: number
+}
+
+export default function PopUp({ children, rect, keepClearAtBottom }: PopUpProps) {
+    const container = useRef<HTMLElement>(document.createElement('div'))
+    const root = useRef<HTMLElement>(document.getElementById('popup-root'))
+
+    function applyStyleAbove(rect: DOMRect, rootRect: DOMRect, element: HTMLElement) {
+        const offsetY = rootRect.bottom - rect.bottom + rect.height
+        const offsetX = rect.left - rootRect.left
+
+        element.style.bottom = offsetY + 'px'
+        element.style.top = ''
+        element.style.left = offsetX + 'px'
+        element.style.width = rect.width + 'px'
+    }
+
+    function applyStyleBelow(rect: DOMRect, rootRect: DOMRect, element: HTMLElement) {
+        const offsetY = rect.bottom - rootRect.top
+        const offsetX = rect.left - rootRect.left
+
+        element.style.top = offsetY + 'px'
+        element.style.bottom = ''
+        element.style.left = offsetX + 'px'
+        element.style.width = rect.width + 'px'
+    }
+
+    useEffect(() => {
+        container.current.className = styles.popup
+        root.current!.appendChild(container.current)
+
+        return () => {
+            root.current!.removeChild(container.current)
+        }
+    }, [container])
+
+    useEffect(() => {
+        const rootRect = root.current!.getBoundingClientRect()
+        const emptySpace = rootRect.bottom - rect.bottom
+
+        emptySpace > keepClearAtBottom
+            ? applyStyleBelow(rect, rootRect, container.current)
+            : applyStyleAbove(rect, rootRect, container.current)
+    }, [rect])
+
+    return createPortal(children, container.current)
+}
